feat: remove no longer needed function_value
In the previous archtecture, the function was automatically treated as a
contextual valuee, even if you really wanted to set a function as a
configuration value. You need to specify a function returning the originally
wanted function to workaround it. Now, the function is no longer treated as a
contextual value unless you explicitly wrap it with DynamicValue, the
function_value wrapper also becomes out of date.